pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paper-radar"
version = "0.1.0"
description = "Daily paper radar for LLM and architecture research."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
  "beautifulsoup4>=4.12.3",
  "feedparser>=6.0.11",
  "jinja2>=3.1.4",
  "lxml>=5.2.2",
  "python-dateutil>=2.9.0",
  "pyyaml>=6.0.2",
  "rapidfuzz>=3.9.6",
  "requests>=2.32.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.2",
]

[project.scripts]
paper-radar = "scripts.main:main"

[tool.setuptools]
py-modules = []

[tool.setuptools.packages.find]
include = ["scripts*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
